Misc changes
Set up a better code structure for handling everything.

Started Deposit function, it's working but won't distribute properly for containers INTO accounts.

Started Create function, works for accounts, but not for containers because those need to be set up more.

Need to set up a function or way to edit self.acctList in the Container class.

Set up self.printContainers which prints the containers and then the accounts for the controller instance.