display working on containers (kinda)
containers now work in the display function, but aren't able to display containers in containers.

created an ItemList class to handle functions for acctList and contList much more effectively in each of their own retrospects.

Got getSum() working for containers! Not 100% sure if it works recursively for containers in containers yet but it should in theory work.