display working!
got it all working properly, no problems!

it was as simple as making COM_INSTANCE only increment when you deposit or withdraw. I also had to change the if statement in get sum from == to <= so that you're adding the totals sum up until the given instance rather than the sum at a singluar point in time.