Save working!
It all works for saving your history and there is a variety of functions in the File class that allow for better organization of code, but there needs to be more done so that the save is actually functional. Such as saving account or container creation. There's also been some precautions implemented so that the user experience doesn't required creating all of these directories and files from scratch!!!