misc
Made a commit on the wrong repo, here's what it had:

id, name, notes

id now displaying saving and loading properly. No use functions for it currently and it isn't put into the addVal functions for all of the money classes.

Name displaying properly
notes still working and saving. Not displaying every yet though.

The other changes I've made are small and to account for me while I'm setting up my own personal budgeting.

Now I can display the balance of all accounts summed up at any point. I added this to the display function so that it automatically displays that all and it makes it easy to check if any of my numbers are wrong or not.

IDs are also generated through an IDGenerator class which is real easy and simple to set up.