Container Percentage rework + misc
Container percentage is now completely reworked to have an editable item selection menu, and a better user interface when selecting items for use in a container's item list.

I also make a yesNo function for questions that are a yes or no so that I don't have to rewrite that code constantly.

I finished the help function (made some small edits to it too) and now it displays everything and doesn't kick the user after a single interaction is asked about. The user can quit it's menu with "quit help"

display now displayes numbers with ending zeros. (.8 is now .80) and aligns numbers with the right side of the column they are in.

Also added in a bool for if the dev wants to increment gb.COM_INSTANCE in autoInputVal. This was intended for some sort of transfer function where the transfer would happen in one instance rather than 2.

Set self.halfRound to be a function for the entire ControllerInstance class rather than just in the save function.