Transfer Function
I made a function to handle transfers between accounts. It is not stored as a transfer and is just stored as two different transactions, so editing (when I get that function done) is not easily preformed.

also renamed inputVal() to transaction()

and IDGen.generate_id() to IDGen.generateID()

autoInputVal() to autoTransaction()

Reworked question (int/float) function. You're now able to set ranges for the function, and if the value is outside of those ranges (there's no limit to how many ranges you can input) then you'll be sent back to try again with another value to put in.