misc + rename func set up
The user can now rename accounts and containers easily.

self.question now will add a "\n" to the string if the dev didn't account for that when writing their questionString. It is designed to only activate if the variable firstInstanceCall is set to True. This way the dev could have no "\n" at the end of their string. This also allows for the function to now run the same code checking for the "\n" every time it is called.