Misc small changes
Added in a function to sort all the instances by their date. Also redid the ID stuff so now it's just a number, that's so much easier to deal with. the temp function also quickly redoes all the IDs so they are right. Fixed a few small bugs, and added in a little display of the most recent transactions in the view menu. Holy I think I forgot how many helper functions I wrote into this program because it was so easy to implement some of the stuff I wanted to do after months and months of not even touching the codebase