misc small fixes
got rid of a naming bug when doing a transaction of any kind and you just press enter instead of entering a name for the transaction. Just needed a little return added to a line.

Also added in some padding and fluff to make it so that we set the id as the last id in controller.history +1. And if that doesn't work (the id of the instance isn't a valid integer) then just default to the length of controller.history as the id and hope and pray that that is the correct number. otherwise idk I'm the dev (and the only user) so who cares rn lol