Few Changes + good temp func in temp.py
added in autosave overwrite feature, and the ability to edit an instances' note.

In the temp.py there's the start of a function to measure the amount spent on food (or any other arbitrary thing you can name consistently) over a duration of time. It's not properly implemented yet though for user experience.