autosave function upgrade
made shit fancy with the autosave function so now you can either view what it's set to, set it to specifically something or just invert it's bool. I also updated the save function.

And the big thing, I made it so that the command function in controller.py doesn't just take one string as a command but splits it into a list so that you can make more complicated commands in one input.